refactor add form
