+++ pages/dashboard/add.vue
@@ -1,36 +1,17 @@
 <script lang="ts" setup>
 import type { FetchError } from 'ofetch'
-import type { Nominatim } from '~/lib/type'
-import { AppFormField } from '#components'
-import { CENTER_MAP } from '~/lib/constants'
-import { InsertLocation } from '~/lib/db/schema/location'
+import type { InsertLocationType } from '~/lib/db/schema/location'
 
 const { $csrfFetch } = useNuxtApp()
-const router = useRouter()
-const submitError = ref('')
 const loading = ref(false)
 const submited = ref(false)
-const mapStore = useMapStore()
-const { handleSubmit, errors, meta, setErrors, setFieldValue, controlledValues } = useForm({
-  validationSchema: toTypedSchema(InsertLocation),
-  initialValues: {
-    long: (CENTER_MAP as [number, number])[0],
-    lat: (CENTER_MAP as [number, number])[1],
-    name: '',
-    description: '',
-  },
-})
+const submitError = ref('')
+const submitErrors = ref<Record<string, string>>({})
 
-function formatNumber(value?: number | undefined) {
-  if (!value) {
-    return 0
-  }
-  return value.toFixed(4)
-}
-
-const onSubmit = handleSubmit(async (values) => {
+async function onSubmit(values: InsertLocationType) {
   try {
     submitError.value = ''
+    submitErrors.value = {}
     loading.value = true
     await $csrfFetch('/api/locations', {
       method: 'post',
@@ -42,53 +23,12 @@ const onSubmit = handleSubmit(async (values) => {
   catch (e) {
     const error = e as FetchError
     if (error.data?.data) {
-      setErrors(error.data?.data)
+      submitErrors.value = error.data?.data
     }
     submitError.value = getFetchErrorMessage(error)
   }
   loading.value = false
-})
-
-function searchResultSelected(result: Nominatim) {
-  setFieldValue('name', result.display_name)
-  mapStore.addedPoint = {
-    description: ',',
-    name: 'added Point',
-    id: 1,
-    long: +result.lon,
-    lat: +result.lat,
-    center_map: true,
-  }
 }
-
-onMounted(() => {
-  mapStore.addedPoint = {
-    description: ',',
-    name: 'added Point',
-    id: 1,
-    long: (CENTER_MAP as [number, number])[0],
-    lat: (CENTER_MAP as [number, number])[1],
-  }
-})
-
-effect(() => {
-  if (mapStore.addedPoint) {
-    setFieldValue('long', mapStore.addedPoint.long)
-    setFieldValue('lat', mapStore.addedPoint.lat)
-  }
-})
-
-onBeforeRouteLeave(() => {
-  if (!submited.value && meta.value.dirty) {
-    // eslint-disable-next-line no-alert
-    const confirm = window.confirm('are u sure u want to leave , unsave Changes will be dissapear')
-    if (!confirm) {
-      return false
-    }
-    mapStore.addedPoint = null
-    return true
-  }
-})
 </script>
 
 <template>
@@ -104,30 +44,6 @@ onBeforeRouteLeave(() => {
     <div v-if="submitError" role="alert" class="alert alert-error">
       <span>{{ submitError }}</span>
     </div>
-    <form class="flex flex-col gap-4" @submit.prevent="onSubmit">
-      <AppFormField name="name" :error="errors.name" label="name" :disabled="loading" />
-      <AppFormField name="description" type="textarea" :error="errors.description" label="description" :disabled="loading" />
-      <p>drag the <Icon name="tabler:map-pin-filled" class="text-warning" /> marker and select location</p>
-      <p class="text-xs text-gray-400">
-        {{ formatNumber(controlledValues?.lat) }} , {{ formatNumber(controlledValues?.long) }}
-      </p>
-
-      <div class="flex justify-end gap-2">
-        <button :disabled="loading" class="btn btn-outline" type="button">
-          <span v-if="loading" class="loading loading-spinner loading-md" />
-
-          <Icon v-else name="tabler:arrow-left" size="24" @click="router.back()" />
-          cancel
-        </button>
-        <button :disabled="loading" class="btn btn-primary" type="submit">
-          <span v-if="loading" class="loading loading-spinner loading-md" />
-
-          <Icon v-else name="tabler:circle-plus-filled" size="24" />
-          submit
-        </button>
-      </div>
-      <div class="divider" />
-    </form>
-    <AppSearch class="mt-4" @result-selected="searchResultSelected" />
+    <LocationForm :on-submit :submited :loading :submit-errors />
   </div>
 </template>
